# Memory improvements to `loadRawSpec`