This PR contains memory usage (and small cpu) improvements to `lawRawSpec` through:
- precomputing the amount of required fixups when building the type graph
- guess-timating the amount of types when reading the btf types reducing the amount of buffer reallocation

Benchmark:
```
>>> sudo go test -run=XXX -bench=BenchmarkParseVmlinux ./internal/btf
```

Before (master branch):
```
goos: linux
goarch: amd64
pkg: github.com/cilium/ebpf/internal/btf
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkParseVmlinux-4   	       9	 118640220 ns/op	96579569 B/op	  786353 allocs/op
PASS
ok  	github.com/cilium/ebpf/internal/btf	2.049s
```

After (updated after review):
```
goos: linux
goarch: amd64
pkg: github.com/cilium/ebpf/internal/btf
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkParseVmlinux-4   	      14	  76856666 ns/op	53361862 B/op	  574838 allocs/op
PASS
ok  	github.com/cilium/ebpf/internal/btf	2.001s
```

Happy to discuss future paths of improvements