@lmb 2022-04-22
It's cool to see these improvements, thanks a lot! I'll do a full review in a bit, but how about merging the non-unsafe commits? I have an idea how we might be able to cut down allocations without pinning the memory, but it needs a bit of experimenting.